move ShowQueryModal to common components
