+++ frontend/components/modals/ShowQueryModal/ShowQueryModal.tsx
