+++ frontend/components/modals/ShowQueryModal/index.ts
